+++ src/lxc/bdev/bdev.c
@@ -59,6 +59,7 @@
 #include "lxclvm.h"
 #include "lxcloop.h"
 #include "lxcoverlay.h"
+#include "lxcrbd.h"
 #include "lxcrsync.h"
 #include "lxczfs.h"
 #include "namespace.h"
@@ -151,6 +152,18 @@ static const struct bdev_ops ovl_ops = {
 	.can_backup = true,
 };
 
+/* rbd */
+static const struct bdev_ops rbd_ops = {
+	.detect = &rbd_detect,
+	.mount = &rbd_mount,
+	.umount = &rbd_umount,
+	.clone_paths = &rbd_clonepaths,
+	.destroy = &rbd_destroy,
+	.create = &rbd_create,
+	.can_snapshot = false,
+	.can_backup = false,
+};
+
 /* zfs */
 static const struct bdev_ops zfs_ops = {
 	.detect = &zfs_detect,
@@ -182,18 +195,6 @@ static int nbd_mount(struct bdev *bdev);
 static int nbd_umount(struct bdev *bdev);
 static bool requires_nbd(const char *path);
 static bool wait_for_partition(const char *path);
-
-/* functions associated with a rdb bdev struct */
-static int rbd_clonepaths(struct bdev *orig, struct bdev *new,
-		const char *oldname, const char *cname,
-		const char *oldpath, const char *lxcpath, int snap,
-		uint64_t newsize, struct lxc_conf *conf);
-static int rbd_create(struct bdev *bdev, const char *dest, const char *n,
-		struct bdev_specs *specs);
-static int rbd_destroy(struct bdev *orig);
-static int rbd_detect(const char *path);
-static int rbd_mount(struct bdev *bdev);
-static int rbd_umount(struct bdev *bdev);
 
 /* helpers */
 /*
@@ -510,167 +511,6 @@ int is_blktype(struct bdev *b)
 		return 1;
 	return 0;
 }
-
-/*
- *   CEPH RBD ops
- */
-
-static int rbd_detect(const char *path)
-{
-	if ( memcmp(path, "/dev/rbd/", 9) == 0)
-		return 1;
-	return 0;
-}
-
-static int rbd_mount(struct bdev *bdev)
-{
-	if (strcmp(bdev->type, "rbd"))
-		return -22;
-	if (!bdev->src || !bdev->dest)
-		return -22;
-
-	if ( !file_exists(bdev->src) ) {
-		// if blkdev does not exist it should be mapped, because it is not persistent on reboot
-		ERROR("Block device %s is not mapped.", bdev->src);
-		return -1;
-	}
-
-	return mount_unknown_fs(bdev->src, bdev->dest, bdev->mntopts);
-}
-
-static int rbd_umount(struct bdev *bdev)
-{
-	if (strcmp(bdev->type, "rbd"))
-		return -22;
-	if (!bdev->src || !bdev->dest)
-		return -22;
-	return umount(bdev->dest);
-}
-
-static int rbd_clonepaths(struct bdev *orig, struct bdev *new,
-		const char *oldname, const char *cname,
-		const char *oldpath, const char *lxcpath, int snap,
-		uint64_t newsize, struct lxc_conf *conf)
-{
-	ERROR("rbd clonepaths not implemented");
-	return -1;
-}
-
-static int rbd_destroy(struct bdev *orig)
-{
-	pid_t pid;
-	char *rbdfullname;
-
-	if ( file_exists(orig->src) ) {
-		if ((pid = fork()) < 0)
-			return -1;
-		if (!pid) {
-			execlp("rbd", "rbd", "unmap" , orig->src, NULL);
-			exit(1);
-		}
-		if (wait_for_pid(pid) < 0)
-			return -1;
-	}
-
-	if ((pid = fork()) < 0)
-		return -1;
-	if (!pid) {
-		rbdfullname = alloca(strlen(orig->src) - 8);
-		strcpy( rbdfullname, &orig->src[9] );
-		execlp("rbd", "rbd", "rm" , rbdfullname, NULL);
-		exit(1);
-	}
-	return wait_for_pid(pid);
-
-}
-
-static int rbd_create(struct bdev *bdev, const char *dest, const char *n,
-		struct bdev_specs *specs)
-{
-	const char *rbdpool, *rbdname = n, *fstype;
-	uint64_t size;
-	int ret, len;
-	char sz[24];
-	pid_t pid;
-
-	if (!specs)
-		return -1;
-
-	rbdpool = specs->rbd.rbdpool;
-	if (!rbdpool)
-		rbdpool = lxc_global_config_value("lxc.bdev.rbd.rbdpool");
-
-	if (specs->rbd.rbdname)
-		rbdname = specs->rbd.rbdname;
-
-	/* source device /dev/rbd/lxc/ctn */
-	len = strlen(rbdpool) + strlen(rbdname) + 11;
-	bdev->src = malloc(len);
-	if (!bdev->src)
-		return -1;
-
-	ret = snprintf(bdev->src, len, "/dev/rbd/%s/%s", rbdpool, rbdname);
-	if (ret < 0 || ret >= len)
-		return -1;
-
-	// fssize is in bytes.
-	size = specs->fssize;
-	if (!size)
-		size = DEFAULT_FS_SIZE;
-
-	// in megabytes for rbd tool
-	ret = snprintf(sz, 24, "%"PRIu64, size / 1024 / 1024 );
-	if (ret < 0 || ret >= 24)
-		exit(1);
-
-	if ((pid = fork()) < 0)
-		return -1;
-	if (!pid) {
-		execlp("rbd", "rbd", "create" , "--pool", rbdpool, rbdname, "--size", sz, NULL);
-		exit(1);
-	}
-	if (wait_for_pid(pid) < 0)
-		return -1;
-
-	if ((pid = fork()) < 0)
-		return -1;
-	if (!pid) {
-		execlp("rbd", "rbd", "map", "--pool", rbdpool, rbdname, NULL);
-		exit(1);
-	}
-	if (wait_for_pid(pid) < 0)
-		return -1;
-
-	fstype = specs->fstype;
-	if (!fstype)
-		fstype = DEFAULT_FSTYPE;
-
-	if (do_mkfs(bdev->src, fstype) < 0) {
-		ERROR("Error creating filesystem type %s on %s", fstype,
-			bdev->src);
-		return -1;
-	}
-	if (!(bdev->dest = strdup(dest)))
-		return -1;
-
-	if (mkdir_p(bdev->dest, 0755) < 0 && errno != EEXIST) {
-		ERROR("Error creating %s", bdev->dest);
-		return -1;
-	}
-
-	return 0;
-}
-
-static const struct bdev_ops rbd_ops = {
-	.detect = &rbd_detect,
-	.mount = &rbd_mount,
-	.umount = &rbd_umount,
-	.clone_paths = &rbd_clonepaths,
-	.destroy = &rbd_destroy,
-	.create = &rbd_create,
-	.can_snapshot = false,
-	.can_backup = false,
-};
 
 //
 // nbd dev ops
